bot/requirements.txt
discord.py==2.7.1
folium==0.20.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
requests==2.31.0
